feat: 용어 도메인과 용어 정보 DTO 조회 영속화 기능 분리
